Align CC scores to SNP positions

Maps TF sample loadings from a CATaNResult to genomic SNP positions using TF ChIP-seq peak BED files. For each CC, every TF's peak regions are assigned that TF's sample loadings, overlapped with the target SNP set, and aggregated by summing scores at each SNP. SNPs not overlapping any peak receive a score of zero.

align_cc_to_snps(
  catan_result,
  peak_dir,
  snp_gr,
  chromosomes = paste0("chr", 1:22),
  population = "custom",
  BPPARAM = BiocParallel::bpparam(),
  verbose = TRUE
)

Arguments

catan_result

A CATaNResult object.

peak_dir

Character. Path to directory containing TF peak BED files (one per TF, named {TF_name}.hg19.bed).

snp_gr

A GRanges object of target SNP positions. Can be created from a .bim file using bim_to_granges.

chromosomes

Character vector of chromosome names to process (default paste0("chr", 1:22)).

population

Character. Label for the reference population (default "custom").

BPPARAM

A BiocParallelParam object for parallel processing (default BiocParallel::bpparam()).

verbose

Logical (default TRUE).

Value

A CATaNAnnotation object with the snp_scores slot populated.

Details

This replicates the logic of the original bedtools-based shell script using GenomicRanges for full R compatibility.

The algorithm proceeds per CC, per chromosome:

Load each TF's peak BED file and assign that TF sample loading as the score.

Combine all TF peaks and find overlaps with the target SNP set (equivalent to bedtools intersect).

For SNPs overlapping multiple TF peaks, sum the scores (equivalent to bedtools merge -d -1 -c 4 -o sum).

Assign score 0 to SNPs with no peak overlap (equivalent to bedtools intersect -v + zero fill).

Concatenate across chromosomes.

Chromosomes are processed in parallel via BiocParallel.

Examples

set.seed(1)
tf <- matrix(sample(0:1, 200, replace = TRUE), 20, 10)
rownames(tf) <- paste0("g", seq_len(20))
colnames(tf) <- paste0("TF", seq_len(10))
ex <- matrix(rnorm(100, 5), 20, 5)
rownames(ex) <- rownames(tf)
colnames(ex) <- paste0("s", seq_len(5))
res <- run_cca(tf, ex, n_cc = 2L, n_hvg = 15L,
               verbose = FALSE)

# Create mock peak BED files
peak_dir <- tempfile("peaks")
dir.create(peak_dir)
for (tf_name in colnames(tf)) {
    bed <- data.frame("chr1", c(100, 500), c(200, 600))
    write.table(bed,
        file.path(peak_dir, paste0(tf_name, ".hg19.bed")),
        sep = "\t", row.names = FALSE,
        col.names = FALSE, quote = FALSE)
}

# Create mock SNPs
library(GenomicRanges)
#> Warning: package ‘GenomicRanges’ was built under R version 4.5.2
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> Loading required package: generics
#> 
#> Attaching package: ‘generics’
#> The following objects are masked from ‘package:base’:
#> 
#>     as.difftime, as.factor, as.ordered, intersect, is.element, setdiff,
#>     setequal, union
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, is.unsorted, lapply,
#>     mapply, match, mget, order, paste, pmax, pmax.int, pmin, pmin.int,
#>     rank, rbind, rownames, sapply, saveRDS, table, tapply, unique,
#>     unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> Warning: package ‘S4Vectors’ was built under R version 4.5.3
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: Seqinfo
snps <- GRanges("chr1",
    IRanges::IRanges(c(150, 550, 9999), width = 1))
annot <- align_cc_to_snps(res, peak_dir, snps,
    chromosomes = "chr1",
    BPPARAM = BiocParallel::SerialParam(),
    verbose = FALSE)